DEFM_Backend/requirements.txt
aiofiles==25.1.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
black==26.1.0
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
docstring-to-markdown==0.17
email-validator==2.3.0
enum34==1.1.10
fastapi==0.124.0
Flask==3.1.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
libewf-python==20240506
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
ninja==1.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pathspec==1.0.3
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.14.0
python-multipart==0.0.21
python-ntfs==0.1
pytokens==0.4.0
pytsk3==20250801
pytz==2025.2
reportlab==4.4.7
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uvicorn==0.40.0
Werkzeug==3.1.4
wheel==0.45.1
zipp==3.23.0
